moving sales files to sales folder
