+++ cypress/e2e/discounts/sales/salesForProducts.js
@@ -1,22 +1,28 @@
 /// <reference types="cypress"/>
-/// <reference types="../../support"/>
+/// <reference types="../../../support"/>
 
 import faker from "faker";
 
-import * as channelsUtils from "../../support/api/utils/channelsUtils";
-import { deleteSalesStartsWith } from "../../support/api/utils/discounts/salesUtils";
-import * as productsUtils from "../../support/api/utils/products/productsUtils";
+import { urlList } from "../../../fixtures/urlList";
+import { createChannel } from "../../../support/api/requests/Channels";
+import { updateChannelInProduct } from "../../../support/api/requests/Product";
+import * as channelsUtils from "../../../support/api/utils/channelsUtils";
+import { deleteSalesStartsWith } from "../../../support/api/utils/discounts/salesUtils";
+import * as productsUtils from "../../../support/api/utils/products/productsUtils";
 import {
   createShipping,
   deleteShippingStartsWith,
-} from "../../support/api/utils/shippingUtils";
+} from "../../../support/api/utils/shippingUtils";
+import { getProductPrice } from "../../../support/api/utils/storeFront/storeFrontProductUtils";
 import {
-  createSaleWithNewVariant,
+  assignProducts,
+  createSale,
+  createSaleWithNewProduct,
   discountOptions,
-} from "../../support/pages/discounts/salesPage";
+} from "../../../support/pages/discounts/salesPage";
 
-xdescribe("Sales discounts for variant", () => {
-  const startsWith = "SalesVar-";
+xdescribe("Sales discounts for products", () => {
+  const startsWith = "SalesProd-";
 
   let productType;
   let attribute;
@@ -26,6 +32,7 @@ xdescribe("Sales discounts for variant", () => {
 
   before(() => {
     cy.clearSessionData().loginUserViaRequest();
+    channelsUtils.deleteChannelsStartsWith(startsWith);
     deleteSalesStartsWith(startsWith);
     productsUtils.deleteProductsStartsWith(startsWith);
     deleteShippingStartsWith(startsWith);
@@ -75,7 +82,7 @@ xdescribe("Sales discounts for variant", () => {
       const discountValue = 50;
       const productPrice = 100;
 
-      createSaleWithNewVariant({
+      createSaleWithNewProduct({
         name: saleName,
         channel: defaultChannel,
         warehouseId: warehouse.id,
@@ -85,10 +92,9 @@ xdescribe("Sales discounts for variant", () => {
         price: productPrice,
         discountOption: discountOptions.PERCENTAGE,
         discountValue,
-      }).then(({ pricing }) => {
-        const priceInResponse = pricing.price.gross.amount;
+      }).then(price => {
         const expectedPrice = (productPrice * discountValue) / 100;
-        expect(expectedPrice).to.be.eq(priceInResponse);
+        expect(expectedPrice).to.be.eq(price);
       });
     },
   );
@@ -101,7 +107,7 @@ xdescribe("Sales discounts for variant", () => {
       const discountValue = 50;
       const productPrice = 100;
 
-      createSaleWithNewVariant({
+      createSaleWithNewProduct({
         name: saleName,
         channel: defaultChannel,
         warehouseId: warehouse.id,
@@ -111,11 +117,60 @@ xdescribe("Sales discounts for variant", () => {
         price: productPrice,
         discountOption: discountOptions.FIXED,
         discountValue,
-      }).then(({ pricing }) => {
-        const priceInResponse = pricing.price.gross.amount;
+      }).then(price => {
         const expectedPrice = productPrice - discountValue;
-        expect(expectedPrice).to.be.eq(priceInResponse);
+        expect(expectedPrice).to.be.eq(price);
       });
+    },
+  );
+
+  it(
+    "should not displayed discount not assign to channel",
+    { tags: ["@sales", "@allEnv"] },
+    () => {
+      const saleName = `${startsWith}${faker.datatype.number()}`;
+      let channel;
+      let product;
+      const discountValue = 50;
+      const productPrice = 100;
+
+      createChannel({ name: saleName }).then(
+        channelResp => (channel = channelResp),
+      );
+      productsUtils
+        .createProductInChannel({
+          name: saleName,
+          channelId: defaultChannel.id,
+          warehouseId: warehouse.id,
+          productTypeId: productType.id,
+          attributeId: attribute.id,
+          categoryId: category.id,
+          price: productPrice,
+        })
+        .then(({ product: productResp }) => {
+          product = productResp;
+          updateChannelInProduct({
+            productId: product.id,
+            channelId: channel.id,
+          });
+        })
+        .then(() => {
+          /* Uncomment after fixing SALEOR-3367 bug 
+           cy.clearSessionData()
+          .loginUserViaRequest("auth", ONE_PERMISSION_USERS.discount) 
+          */
+
+          cy.visit(urlList.sales);
+          cy.expectSkeletonIsVisible();
+          createSale({
+            saleName,
+            channelName: channel.name,
+            discountValue,
+          });
+          assignProducts(product.name);
+          getProductPrice(product.id, defaultChannel.slug);
+        })
+        .then(price => expect(price).to.equal(productPrice));
     },
   );
 });
